requirements.txt
fastapi
uvicorn
./openclaw_vendor
playwright
httpx
pydantic
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
cryptography
pyyaml
